Revert "Merge pull request #11 from d0k3/master"
This reverts commit a20fd5ba70dc5b64128e989e113a513a4cf07725, reversing
changes made to 81835439e189371a75de7e75bc3860adfcb27ad8.

---

This experimental feature was supposed to be conditionally enabled to take end-users with slow SD cards into account, and was supposed to be used with A9NC (which, anyways, is extremely niche).

It was certainly **NOT** designed to be used merely as an inappropriate and lazy way to chainload a payload from SD card or NAND (doing so is also extremely inefficient due to the need to write a file and reboot).

People concerned should see b9s's own source code to have see a proper example to load FIRM payloads correctly (I could help anyone who would have trouble integrating that code into their projects) instead.

---

Conflicts:
	stage2/arm9/source/fs.c
	stage2/arm9/source/fs.h
	stage2/arm9/source/main.c